Remove Chembl extraction re-export from application
