+++ src/bioetl/application/services/__init__.py
@@ -1,5 +1,3 @@
 """Application services - orchestration layer."""
 
-from bioetl.application.services.chembl_extraction import ChemblExtractionServiceImpl
-
-__all__ = ["ChemblExtractionServiceImpl"]
+__all__: list[str] = []
