+++ src/bioetl/application/services/chembl_extraction.py
@@ -1,12 +1,3 @@
-"""
-Application-level entrypoint for ChEMBL extraction service.
+"""Application-level ChEMBL extraction service entrypoint placeholder."""
 
-Re-exports the infrastructure implementation to keep backward-compatible
-imports for pipelines and tests.
-"""
-
-from bioetl.infrastructure.clients.chembl.impl.chembl_extraction_service_impl import (
-    ChemblExtractionServiceImpl,
-)
-
-__all__ = ["ChemblExtractionServiceImpl"]
+# ChemblExtractionServiceImpl is no longer exposed at the application layer.
